requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
qdrant-client
faiss-cpu
numpy
scikit-learn
joblib
sentence-transformers
torch
regex
streamlit
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
pytest
